==> climate/scripts_ds/1. data_preprocessing.ipynb <==
### Import libraries

import pandas as pd
import numpy as np
import re

pd.set_option('display.max_columns', None)  # None means unlimited columns

### Read datasets

judet = "timisoara"

df = pd.read_excel(f"../data/data_tabular/raport_{judet}.xls", skiprows=7, header=[0, 1, 3])

### Preprocess dataframe

# Save data in a temp df
data = df.set_index('Luna/Zi').reset_index()[['Luna/Zi']]

# Drop it from the original df
df.drop(['Luna/Zi'], axis=1, inplace=True)

# Reshape the date column in a proper format
def tuple_to_datetime(t):
    year_month, day = t[0].split("-")
    day = day.strip(',')
    return pd.to_datetime(f"{year_month}-{day}")

# Apply the function to the 'Luna/Zi' column
data['Luna/Zi'] = data['Luna/Zi'].apply(tuple_to_datetime)
data['Luna/Zi'] = pd.to_datetime(data['Luna/Zi'], format='%Y%m-%d')

# Set the index as date:
df.index = data['Luna/Zi']

# Merge level 1 and level 2 of our dataframe
def merge_levels(df, levels_to_merge):
    new_levels = df.columns.to_frame()
    new_levels['merged'] = new_levels[levels_to_merge].apply(lambda row: '_'.join(row), axis=1)
    
    for level in levels_to_merge:
        new_levels.drop(level, axis=1, inplace=True)
    
    df.columns = pd.MultiIndex.from_frame(new_levels)
    return df

df = merge_levels(df, [1, 2])

# Stack the dataframe in order to keep only the variables as columns, but date and region as indexes
df = df.stack(0)

# Set the proper index
df = df.reset_index()
df.rename({0:'County'}, axis=1, inplace=True)

# Convert all the columns to numeric dtype and replace with NaN if any special character's been found
for col in df.columns[2:]:
    df[col] = pd.to_numeric(df[col], errors='coerce')

# Extract the letter corresponding to each county
def extract_letters(s):
    match = re.search(r'^([a-zA-Z]+)', s)
    if match:
        return match.group(1)
    else:
        return None
    
county = extract_letters(df['County'][0])
df.drop('County', axis=1, inplace=True)

# Compute the average for each station from the same County to handle the missing values
def average_non_nan(group):
    return group.apply(lambda x: np.sum(x) / np.count_nonzero(~np.isnan(x)) if np.count_nonzero(~np.isnan(x)) else np.nan)

# Group by 'Luna/Zi' and apply the custom function to each group
df = df.groupby('Luna/Zi').apply(average_non_nan)
df.insert(loc=0, column='County', value=county)

df.to_excel(f"../data/data_tabular/{judet}_preprocessed.xlsx")
