==> climate/scripts_ds/2. data_concatenation.ipynb <==
### Import libraries

import pandas as pd
import os

### Read dataframes

file_names = ['bucuresti_preprocessed.xlsx', 
              'iasi_preprocessed.xlsx', 
              'cluj_preprocessed.xlsx', 
              'dolj_preprocessed.xlsx',
              'timisoara_preprocessed.xlsx',
              'constanta_preprocessed.xlsx',
              'brasov_preprocessed.xlsx']

path_to_files = '../data/data_tabular/'  # replace with your directory path
dfs = {file: pd.read_excel(os.path.join(path_to_files, file)) for file in file_names}

### Concatenate dataframes

df = pd.concat(dfs.values(), axis=0, ignore_index=True, sort=False)

### Export dataframe

df.to_excel('../data/data_tabular/counties_preprocessed.xlsx', index=False)
