==> climate/scripts_ds/4. data_transformation.ipynb <==
import pandas as pd
import numpy as np
import ast 
import json
import re

#### Convert string to list and remove duplicates

df_phen = pd.read_csv('../data/datasets/dataset_merged.csv')

def remove_duplicates_from_str_list(str_list):
    # Convert string representation of list to actual list
    actual_list = ast.literal_eval(str_list)
    # Return list without duplicates while maintaining order
    cleaned_list = []
    for item in actual_list:
        if item not in cleaned_list:
            cleaned_list.append(item)
    return cleaned_list

df_phen['phenomena'] = df_phen['phenomena'].apply(remove_duplicates_from_str_list)

#### Function to remove duplicates from a list

def remove_duplicates_from_actual_list(lst):
    # Use dict.fromkeys() to preserve order and remove duplicates
    return list(dict.fromkeys(lst))

df_concatenated_levels['phenomena'] = df_concatenated_levels['phenomena'].apply(remove_duplicates_from_actual_list)
df_concatenated_levels['phenomena'] = df_concatenated_levels['phenomena'].apply(lambda x: [item for item in x if item != 'Avalanșă'])
df_concatenated_levels.to_excel('data/set_complet_precip.xlsx')

##### Compute the daily average of the water levels

#### Re-evaluate the dataset and maping phenomena

df_phen = pd.read_excel('../data/datasets/corrected_dataset.xlsx')

df_phen['phenomena'] = df_phen['phenomena'].apply(ast.literal_eval)

word_mapping = {
    'Ninsori': 'Ninsoare',
    'Ploi': 'Ploaie',
    'Vant puternic': 'Vant',
    'Vânt puternic': 'Vant',
    'Vant mai activ': 'Vant',
    'Ploi locale': 'Ploaie',
    'Descărcare electrică':'Fulger',
    'Descărcări electrice': 'Fulger',
    'Vijelii': 'Vijelie',
    'Furtuni': 'Furtuna',
    'Vant intens': 'Vant',
    'Descarcari electrice': 'Fulger',
    'Intensificări ale vântului': 'Vant',
    'Căderi de grindină': 'Grindina',
    'Grindină': 'Grindina',
    'Fenomene eletrice': 'Fulger', 
    'Ploi abundente': 'Ploaie',
    'Ploi consistente': 'Ploaie',
    'Ploi torentiale': 'Ploaie',
    'Frecvente descărcări electrice': 'Fulger',
    'Intensificări de scurtă durată ale vântului': 'Vant',
    'Vânt intensificat': 'Vant',
    'Ploi slabe': 'Ploaie',
    'Ploi de scurta durata': 'Ploaie',
    'Frecvente descarcari electrice': 'Fulger',
    'Intensificari ale vantului cu aspect de vijelie': 'Vijelie',
    'Caderi de grindina': 'Grindina',
    'Caniculă': 'Canicula',
    'Vânt intens': 'Vant',
    'Căldură insuportabilă': 'Canicula',
    'Fulgere' : 'Fulger',
    'Intensificari de scurta durata ale vantului cu aspect de vijelie': 'Vijelie',
    'Izolat grindină':'Grindina',
    'Vânt activ': 'Vant',
    'Averse locale': 'Averse',
    'Averse zdravene': 'Averse',
    'Vânt tare în zonele de munte': 'Vant',
    'Averse torențiale': 'Averse',
    'Strat consistent de zapada': 'Zapada',
    'Strat consistent de zăpadă': 'Zapada',
    'Ninsori': 'Ninsoare',
    'Ploi': 'Ploaie',
    'Ceața': 'Ceata',
    'Fulguială': 'Ninsoare',
    'Vânt': 'Vant',
    'Câțiva stropi': 'Ploaie',
    'Ninsori mixte': 'Ninsoare',
    'Ceată': 'Ceata',
    'Ceață': 'Ceata',
    'Intensificări de vânt': 'Vant',
    'Fenomene electrice': 'Fulger',
    'Fulgi de zăpadă': 'Ninsoare',
    'Fulguiei': 'Ninsoare',
    'Precipitații mixte': 'Lapovita',
    'Lapoviță': 'Lapovita',
    'Ploaie slabă': 'Ploaie',
    'Aversă': 'Averse',
    'Averse cu descărcări electrice': 'Averse',
    'Averse de ploaie': 'Averse',
    'Căderi de ploaie': 'Ploaie',
    'Rabufniri ale vântului': 'Vant',
    'Posibil chiar vijelii': 'Vijelie',
    'Porniri ale vantului': 'Vant',
    'Vânt cu aspect de vijelie': 'Vijelie',
    'Averse consistente': 'Averse',
    'Temporar va ploua': 'Ploaie',
    'Canitati importante de apa': 'Ploaie',
    'Cantitati mai insemnate de apa': 'Ploaie',
    'Canitati insemnate de apa': 'Ploaie',
    'Vant in rafale':'Vant',
    'Temperaturi ridicate':'Canicula',
    'Posibile ploi': 'Ploaie',
    'Ploi temporare': 'Ploaie',
    'Ploi mixte cu lapoviță și polei': 'Lapovita',
    'Ninsori mai linistite': 'Ninsoare',
    'Ninsori abia in noaptea de sambata spre duminica': 'Ninsoare',
    'Ninsori de-a lungul zilei': 'Ninsoare',
    'Ninsori locale': 'Ninsoare',
    'Ninsori troienite': 'Viscol',
    'Ninsori viscolite': 'Viscol',
    'Gero' : 'Ger',
    'Ninsoare mixtă':'Ninsoare',
    'Fulgi': 'Ninsoare',
    'Fulguieli': 'Ninsoare',
    'Ploi de vară': 'Ploaie',
    'Răbufniri ale vântului': 'Vant',
    'Porniri ale vântului': 'Vant',
    'Ploaie torențială': 'Ploaie',
    'Ploi torentiale': 'Ploaie',
    'Cantități mari de apă': 'Ploaie',
    'Precipitații slabe mixte': 'Lapovita',
    'Vânt slab': 'Vant',
    'Intensificări de scurtă durată': 'Vant',
    'Ploi insemnate cantitativ': 'Ploaie',
    'Ploi însemnate cantitaiv': 'Ploaie',
    'Averse cu fenomene electrice': 'Fulger',
    'Vreme foarte caldă': 'Canicula',
    'Atmosferă înăbușitoare': 'Canicula',
    'Ninsori abundente': 'Ninsori',
    'Mixt': 'Lapovita',
    'Precipitații slabe mixte': 'Lapovita',
    'ITU critic de 80': 'Canicula',
    'Posibile condiții de grindină': 'Grindina',
    'Grindină': 'Grindina',
    'Aversă torențială': 'Averse',
    'Precipitatii mixte': 'Lapovita',
    'Precipitatii salbe mixte': 'Lapovita',
    'Depuneri de polei': 'Polei',
    'Averse torentiale': 'Averse',
    'Ploi cu caracter de aversă': 'Ploaie',
    'Vant moderat':' Vant',
    'Vant nervos': 'Vant',
    'Ploi slabe cantitativ': 'Ploaie',
    'Ploi moderate cantitativ': 'Ploaie',
    'Posibil grindină': 'Grindina',
    'Rafale de vânt': 'Vant',
    'Intensificari ale vantului': 'Vant',
    'Vânt moderat': 'Vant',
    'Vânt scurt intensificat': 'Vant',
    'Temperaturi maxime record': 'Canicula',
    'Ploi zgomotoase': 'Ploaie',
    'Averse cu fulgere și vânt': 'Averse',
    'Intensificari de scurta durata ale vantului': 'Vant',
    'Izolat grindina':'Grindina',
    'Vant activ':'Vant',
    'Ploi rapide de vara':'Ploaie',
    'Ploi torențiale': 'Ploaie',
    'Ploi rapide de vară': 'Ploaie',
    'Ploi trecatoare': 'Ploaie',
    'Vânt moderat': 'Vant',
    'Cantități mai mari de apă': 'Ploaie',
    'Probabil grindină': 'Grindina',
    'Condiții de ploaie': 'Ploaie',
    'Fenomene de polei': 'Polei',
    'Precipitații sub formă de lapoviță și ninsoare': 'Lapovita',
    'Ploi pe suprafețe restrânse': 'Ploaie',
    'Picură din loc în loc': 'Ploaie',
    'Precipitații slabe, mixte': 'Lapovita',
    'Ploi izolate': 'Ploaie',
    'Ploi peste medie': 'Ploaie',
    'Averse slabe': 'Averse',
    'Vant cu rabufniri': 'Vant',
    'Ploi in zona de munte': 'Ploaie',
    'Ploi de vara':'Ploaie',
    'Averse isotite':'Averse',
    'Averse sporadice':'Averse',
    'Averse cu fulgere, vânt și grindină':'Averse',
    'Averse posibil torentiale':'Averse',
    'Ploi razlete':'Ploaie',
    'Ploi de scurtă durată':'Ploaie',
    'Ploi locale și temporare':'Ploaie',
    'Ploi serioase':'Ploaie',
    'Ploi sporadice':'Ploaie',
    'Averse locale spre seară':'Averse',
    'Formare de polei':'Polei',
    'Ninsori sporadice':'Ninsoare',
    'Ninsori slabe':'Ninsoare',
    'Ninsoare slabă':'Ninsoare',
    'Vant rece':'Vant',
    'Ninsoare slabă':'Ninsoare',
    'Ninsori slabe':'Ninsoare',
    'Fulguieli izolate':'Ninsoare',
    'Vant mai intens':'Vant',
    'Ninsoare slabă': 'Ninsoare',
    'Burniță/Fulguială': 'Burnita',
    'Burniță': 'Burnita',
    'Fulgi de zapada': 'Ninsoare',
    'Ninsori și lapoviță în seara și noaptea de vineri spre sâmbătă': 'Ninsoare',
    'Șanse reduse de ploaie': 'Ploaie',
    'Ploi în mai multe reprize': 'Ploaie',
    'Vant intensificat': 'Vant',
    'Valuri uriase': 'Valuri',
    'Descarcări electrice': 'Fulger',
    'Intensificări ale vântului cu aspect de vijelie': 'Vant',
    'Caldură insuportabilă': 'Canicula',
    'Disconfort termic foarte mare': 'Canicula',
    'Ploi trecătoare': 'Ploaie',
    'Disconfort termic ridicat': 'Canicula',
    'Caldură': 'Caldura',
    'Căldură': 'Caldura',
    'Temperaturi record': 'Canicula',
    'Averse cu descarcari electrice': 'Fulger',
    'Vant mai nervos': 'Vant',
    'Cod galben de caniculă':'Canicula',
    'Cod portocaliu de caniculă': 'Canicula',
    'Intensificări temporare ale vântului': 'Vant',
    'Ploi locale în Muntii Banatului': 'Ploaie',
    'Pericol de ploi moderate cantitativ': 'Ploaie',
    'Cod portocaliu: Vânt puternic și zăpadă troienită': 'Viscol',
    'Cod galben: Intensificări ale vântului și zăpadă spulberată':'Viscol',
    'Cod galben: Intensificări ale vântului și ninsori moderate cantitativ și viscolite':'Viscol',
    'Ploaie cu lapoviță': 'Lapovita',
    'Ninsori moderate cantitativ': 'Ninsoare',
    'Ceată densă': 'Ceata',
    'Cod galben de vânt intens': 'Vijelie',
    'Vânt mai insistent': 'Vant',
    'Izolat cu aspect de vijelie':'Vijelie',
    'Ploi de scurtă durată': 'Ploaie',
    'Rabufniri de vânt': 'Vant',
    'Rabufniri ale vantului': 'Vant',
    'Izolat de grindină': 'Grindina',
    'Cod portocaliu de inundatii': 'Inundatie',
    'Inundație': 'Inundatie',
    'Cod galben de inundatie': 'Inundatie',
    'Ploi mai zgomotoase': 'Ploaie',
    'Căldură toropitoare': 'Canicula',
    'Intensificari de vant': 'Vant',
    'Vânt insistent': 'Vant',
    'Ploi mixte cu polei': 'Polei',
    'Burnită': 'Burnita',
    'Burniță':'Burnita',
    'Burnița': 'Burnita',
    'Vant insistent': 'Vant',
    'Intensificări de scurtă durată ale vântului cu aspect de vijelie': 'Vijelie',
    'Intensificări puternice ale vântului cu aspect de vijelie': 'Vijelie',
    'Inundație': 'Inundatie',
    'Posibil vijelii': 'Vijelie',
    'Însorit': 'Vreme buna',
    'Vreme frumoasă': 'Vreme buna',
    'Vreme calda': 'Vreme buna',
    'Vreme caldă': 'Vreme buna',
    'Temperaturi în creștere': 'Vreme buna',
    'Averse cu fulgere': 'Fulger',
    'Stres termic': 'Canicula',
    'Racoare': 'Vreme rece',
    'Vânt nervos': 'Vant',
    'Răcire accentuată': 'Vreme rece',
    'Fulguieri': 'Ninsoare',
    'Înnorat': 'Innorat',
    'Temperaturi modeste': 'Vreme rece',
    'Vreme instabilă': 'Precipitatii',
    'Instabilitate atmosferică': 'Instabilitate atmosferica',
    'Viitură': 'Inundatie',
    'Averse cu tunete și fulgere': 'Fulger',
    'Averse trecatoare': 'Averse',
    'Mixtă': 'Lapovita',
    'Innorări persistente': 'Innorari',
    'Vreme insorită': 'Vreme buna',
    'Disconfort termic accentuat': 'Canicula',
    'Innorări trecătoare': 'Innorari',
    'Cer noros': 'Innorari',
    'Cer acoperit': 'Innorari', 
    'Vreme placuta': 'Vreme buna',
    'Frecvente descarcări electrice': 'Fulger',
    'Scădere temperaturi': 'Vreme rece',
    'Innorari temporare': 'Innorari',
    'Vreme plăcută': 'Vreme buna',
    'Izolat vijelii': 'Vijelie',
    'Instabilitate atmosferică temporar accentuată': 'Instabilitate atmosferica',
    'Vreme inchisa': 'Innorari',
    'Zăpadă spulberată': 'Viscol',
    'Temperaturi negative': 'Vreme rece',
    'Innorări temporare': 'Innorari',
    'Încălzire': 'Vreme buna',
    'Precipitații moderate': 'Precipitatii',
    'Precipitații excedentare': 'Precipitatii',
    'Vreme frumoasa':'Vreme buna',
    'Înnorări': 'Innorari',
    'Caldura': 'Canicula',
    'Ghețuș': 'Polei',
    'Vreme insorita':'Vreme buna',
    'Caderi de grindină': 'Grindina',
    'Vreme bună': 'Vreme buna',
    'Vreme insorita': 'Vreme buna',
    'Instabilitate atmosferică accentuată': 'Instabilitate atmosferica',
    'Incalzire': 'Vreme buna',
    'Racire': 'Vreme rece',
    'Zapada': 'Ninsoare',
    'Temperaturi scăzute': 'Vreme rece',
    'Vreme caldă': 'Vreme buna',
    'Zăpadă': 'Ninsoare',
    'Nori': 'Innorari',
    'Frig': 'Vreme rece',
    'Innorări': 'Innorari',
    'Disconfort termic': 'Canicula',
    'Precipitații slabe': 'Precipitatii',
    'Soare': 'Vreme buna',
    'Precipitații': 'Precipitatii',
    'Instabilitate atmosferică': 'Instabilitate atmosferica',
    'Vreme frumoasă': 'Vreme buna',
    'Innorat': 'Innorari',
    'Răcire': 'Vreme rece',
    'Instabilitate': 'Instabilitate atmosferica',
    'Negură': 'Ceata',
    'Gheață': 'Polei',
    'instabilitate atmosferica': 'Instabilitate atmosferica'
}



def process_word_list(phenomena_list):
    # Map words using the word_mapping dictionary
    processed_list = [word_mapping.get(word, word) for word in phenomena_list]
    # Remove duplicates while preserving order
    return list(dict.fromkeys(processed_list))

# Apply the function to the phenomena column
df_phen['phenomena'] = df_phen['phenomena'].apply(process_word_list)

### Get rid of the unnecessary phenomena

data_exploded = df_phen.copy()
data_exploded = data_exploded.explode('phenomena')

# Counting occurrences for each phenomenon
phenomena_counts = data_exploded['phenomena'].value_counts()

# Identifying phenomena with less than 10 occurrences
rare_phenomena = phenomena_counts[phenomena_counts < 10].index

# Filtering out the rare phenomena from the exploded dataframe
data_filtered = data_exploded[~data_exploded['phenomena'].isin(rare_phenomena)]

data_filtered['phenomena'].value_counts()

# List of phenomena to be removed
remove_phenomena = ['Uragan', 'Alunecare', 'Cod galben', 'Cod portocaliu', 'Eroziune', 'Cutremur']

# Filtering the exploded dataset
filtered_data_remove_outliers = data_filtered[~data_filtered['phenomena'].isin(remove_phenomena)]

# Checking the updated dataset
filtered_data_remove_outliers['phenomena'].value_counts()

df_phen = filtered_data_remove_outliers.groupby(['date', 'region', 'county', 'pm10_quality', 'air_pressure', 'temperature',
                                      'humidity']
                                     ).agg({'phenomena': list}).reset_index()

### Translate every phenomenon in English

weather_translation = {
    'Ninsoare': 'Snow',
    'Vant': 'Wind',
    'Ploaie': 'Rain',
    'Lapovita': 'Sleet',
    'Inundatie': 'Flood',
    'Ceata': 'Fog',
    'Ger': 'Frost',
    'Viscol': 'Blizzard',
    'Polei': 'Freezing rain',
    'Fulger': 'Lightning',
    'Innorari': 'Cloudiness',
    'Vreme rece': 'Cold weather',
    'Burnita': 'Drizzle',
    'Cer variabil': 'Variable sky',
    'Cer senin': 'Clear sky',
    'Vreme buna': 'Good weather',
    'Precipitatii': 'Precipitation',
    'Vijelie': 'Whirlwind',
    'Canicula': 'Heatwave',
    'Averse': 'Showers',
    'Furtună': 'Storm',
    'Grindina': 'Hail',
    'Tunete': 'Thunder',
    'Secetă': 'Drought',
    'Vreme schimbătoare': 'Changeable weather',
    'Instabilitate atmosferica': 'Atmospheric instability',
    'Ciclon': 'Cyclone',
    'Chiciură': 'Rime',
    'Trăsnet': 'Thunderbolt',
    'Tunet': 'Thunder',
    'Brumă': 'Frost',
    'Incendii': 'Fires',
    'Ploaie înghețată': 'Frozen rain',
    'Tornadă': 'Tornado',
    'Frozen rain': 'Freezing rain',
    'Thunderbolt': 'Lightning', 
    'Thunder': 'Lightning', 
    'Changeable weather':'Atmospheric instability'
}

def translate_word_list(phenomena_list):
    # Map words using the weather_translation dictionary
    processed_list = [weather_translation.get(word, word) for word in phenomena_list]
    # Remove duplicates while preserving order
    return list(dict.fromkeys(processed_list))

# Apply the function to the phenomena column
df_phen['phenomena'] = df_phen['phenomena'].apply(translate_word_list)

df_phen['date'] = pd.to_datetime(df_phen['date'])
df_phen.sort_values(by='date', inplace=True)
df_phen.to_excel('../data/datasets/final.xlsx', index=False)
